internal/gamepad: eliminate Cgo on macOS using purego/objc
Rewrite api_gc_darwin.go and haptics_gc_darwin.go to use purego/objc
instead of Cgo for GameController and CoreHaptics framework access.
This removes the last Cgo dependency in the gamepad package on macOS.

Updates #1162

+++ internal/gamepad/gamepad_gc_darwin.go
@@ -37,7 +37,7 @@ func (*nativeGamepadsGC) update(gamepads *gamepads) error {
 
 type nativeGamepadGC struct {
 	controller           uintptr
-	buttonMask           uint16
+	buttonMask           uint32
 	hasDualshockTouchpad bool
 	hasXboxPaddles       bool
 	hasXboxShareButton   bool
